requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Jinja2==3.1.4
Werkzeug==3.0.3
itsdangerous==2.2.0
gunicorn==21.2.0
google-generativeai>=0.8.0
python-dotenv==1.0.0
